+++ CHANGELOG.md
@@ -5,7 +5,17 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0]
+## [Unreleased]
+
+### Added
+
+- Added `flashCookieOptions` object that can be exported from `sveltekit-flash-message/server` and used to set cookie options.
+
+### Security
+
+- Default cookie option for `sameSite` is set to `strict`.
+
+## [2.0.0] - 2023-07-10
 
 ### Added